OZ-349: Custom exceptions for authentication and demo data generation
