+++ src/main/java/com/ozonehis/ozone_demo_data/exceptions/AuthenticationException.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright © 2025, Ozone HIS <info@ozone-his.com>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.ozonehis.ozone_demo_data.exceptions;
+
+public class AuthenticationException extends RuntimeException {
+
+    public AuthenticationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
